test: Split test files
